refactor(frontend): move ActiveProjectProvider to LayoutContent
Instead of a new ProtectedLayoutClient.tsx file, move
ActiveProjectProvider up into LayoutContent where the server-fetched
initialActiveProject prop is already available. The root layout fetches
the active project server-side (same pattern as org name) and passes it
down so the provider is seeded before the first paint.

(protected)/layout.tsx stays a plain client component; no extra file
needed.
